perf(worker): split fetch handler module
